+++ index.ts
@@ -4,7 +4,7 @@ import { CONSTANTS, parseParams } from "./src/utils.ts";
 import { COLORS, Theme } from "./src/theme.ts";
 import { Error400, Error404 } from "./src/error_page.ts";
 import "https://deno.land/x/dotenv@v0.5.0/load.ts";
-import {staticRenderRegeneration} from "./src/static_render_regeneration.ts";
+import {staticRenderRegeneration} from "./src/StaticRenderRegeneration/index.ts";
 const apiEndpoint = Deno.env.get("GITHUB_API") || CONSTANTS.DEFAULT_GITHUB_API;
 const client = new GithubAPIClient(apiEndpoint);
